Adjusted Database User and Packs
I had to remodel the database so when we remove an Admin it would break the link with the packs created and it could turn to be a customer to. So I united those table in a single type with a role.

I'm adjusting the links with packs in this commit too