Webhook + Subscription + VIP Group Chat
Now SWBot receive a webhook from the Checkout to understand who bought a subscription and send to them the link to enter in the VIP Group.

Also now is possible to call the bot in a group or channel and turn that into a VIP Group.

For a better future the code that receive the webhook from the checkout shoud be isolated and called so it'll be easier to do maintance