Changed bots name and Modified subscriptionBought
Changed bots name to follow the pattern of others bot and work normally with the dynamic db. Isolated subscriptionBought into subscription controller for better future maintance. Altought it isn't perfect and still need to do maintance in the main bot file which is not really good if we have a tons of bots